#509 - mu: reorganize core
