+++ src/mu/core/indirect.rs
@@ -5,22 +5,10 @@
 #![allow(unused_braces)]
 #![allow(clippy::identity_op)]
 use crate::{
-    core::{
-        exception,
-        frame::Frame,
-        mu::Mu,
-        types::{Tag, TagType, Type},
-    },
+    core::types::{Tag, TagType, Type},
     modular_bitfield::specifiers::{B1, B60},
-    types::{
-        fixnum::Fixnum,
-        symbol::{Core as _, Symbol},
-        vecimage::{TypedVec, VecType},
-        vector::Core as _,
-    },
+    types::symbol::{Core as _, Symbol},
 };
-
-use futures::executor::block_on;
 
 // little-endian tag format
 #[derive(Copy, Clone)]
@@ -44,71 +32,19 @@ lazy_static! {
         (Symbol::keyword("t"), Type::T),
         (Symbol::keyword("vector"), Type::Vector),
     ];
-    static ref INFOTYPE: Vec<Tag> = vec![
-        Symbol::keyword("cons"),
-        Symbol::keyword("func"),
-        Symbol::keyword("stream"),
-        Symbol::keyword("struct"),
-        Symbol::keyword("symbol"),
-        Symbol::keyword("vector"),
-    ];
 }
 
 pub trait Core {
-    fn to_type(_: Tag) -> Option<Type>;
-    fn hp_info(_: &Mu) -> (usize, usize);
-    fn hp_type(_: &Mu, _: Type) -> (u8, usize, usize, usize);
+    fn to_indirect_type(_: Tag) -> Option<Type>;
 }
 
-impl Core for Mu {
-    fn to_type(keyword: Tag) -> Option<Type> {
+impl Core for IndirectTag {
+    fn to_indirect_type(keyword: Tag) -> Option<Type> {
         TYPEMAP
             .iter()
             .copied()
             .find(|tab| keyword.eq_(tab.0))
             .map(|tab| tab.1)
-    }
-
-    fn hp_info(mu: &Mu) -> (usize, usize) {
-        let heap_ref = block_on(mu.heap.read());
-
-        (heap_ref.page_size, heap_ref.npages)
-    }
-
-    fn hp_type(mu: &Mu, htype: Type) -> (u8, usize, usize, usize) {
-        let heap_ref = block_on(mu.heap.read());
-        let alloc_ref = block_on(heap_ref.alloc_map.read());
-
-        alloc_ref[htype as usize]
-    }
-}
-
-pub trait MuFunction {
-    fn mu_hp_info(_: &Mu, _: &mut Frame) -> exception::Result<()>;
-}
-
-impl MuFunction for Mu {
-    fn mu_hp_info(mu: &Mu, fp: &mut Frame) -> exception::Result<()> {
-        let (pagesz, npages) = Self::hp_info(mu);
-
-        let mut vec = vec![
-            Symbol::keyword("t"),
-            Fixnum::as_tag((pagesz * npages) as i64),
-            Fixnum::as_tag(npages as i64),
-            Fixnum::as_tag(npages as i64),
-        ];
-
-        for htype in INFOTYPE.iter() {
-            let (_, size, alloc, in_use) = Self::hp_type(mu, Self::to_type(*htype).unwrap());
-
-            vec.push(*htype);
-            vec.push(Fixnum::as_tag(size as i64));
-            vec.push(Fixnum::as_tag(alloc as i64));
-            vec.push(Fixnum::as_tag(in_use as i64));
-        }
-
-        fp.value = TypedVec::<Vec<Tag>> { vec }.vec.to_vector().evict(mu);
-        Ok(())
     }
 }
