+++ src/mu/core/mod.rs
@@ -8,10 +8,13 @@ pub mod compile;
 pub mod direct;
 pub mod exception;
 pub mod frame;
-pub mod functions;
+pub mod funcall;
+pub mod heap;
 pub mod indirect;
 pub mod mu;
 pub mod namespace;
 pub mod reader;
 pub mod readtable;
+pub mod stream;
+pub mod system;
 pub mod types;
